third_party/blink/web_tests/platform/generic/fast/js/inc-bracket-assign-subscript-expected.txt

Test for regression against

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testPreIncBracketAccessWithAssignSubscript() is true
PASS testPostIncBracketAccessWithAssignSubscript() is true
PASS successfullyParsed is true

TEST COMPLETE
